% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{DGP}
\alias{DGP}
\title{Data Generate Process}
\usage{
DGP(n, m, rho)
}
\arguments{
\item{n}{sample size N = n + 1}

\item{m}{dimensions}

\item{rho}{correlation coefficients when i not equal to j}
}
\value{
a n*m matrix
}
\description{
Data Generate Process
}
\examples{
DGP(n = 5, m = 4, rho = 0.1)
}
