% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Wnm_rcpp}
\alias{Wnm_rcpp}
\title{A Rcpp version of Wnm}
\usage{
Wnm_rcpp(X)
}
\arguments{
\item{X}{a n*m matrix}
}
\value{
a vector contain [1]Wnm and [2]p-value
}
\description{
A Rcpp version of Wnm
}
\examples{
Wnm_rcpp(DGP(n = 5, m = 4, rho = 0))
}
